decouple rewriter
